+++ bin2hex/__init__.py
@@ -5,4 +5,4 @@
 # SPDX-License-Identifier: BSD-3-Clause
 #
 
-__version__ = "2.1.0"
+__version__ = "2.2.0"